fix(root): moves __tests__ => tests
