+++ README.md
@@ -1,7 +1,6 @@
 # Data Structures 🦄
 
 ![Typed with TypeScript](https://flat.badgen.net/badge/icon/Typed?icon=typescript&label&labelColor=blue&color=555555)
-[![lerna](https://img.shields.io/badge/maintained%20with-lerna-cc00ff.svg)](https://lerna.js.org/)
 
 #### Basic. Functional. Typed. Data Structures.
 
@@ -20,18 +19,18 @@ Each package contains a data structure
 
 ## Why
 
-#### Improving on data structures is important! Day-to-day, engineering work doesn't deal with data structures. 
+#### Improving on data structures is important! Day-to-day, engineering work doesn't deal with data structures.
 
-This repository attempts to provide the simpliest approach to data structures in JavaScript by making each structure functional and typed, but not too much; very little in fact. 
+This repository attempts to provide the simpliest approach to data structures in JavaScript by making each structure functional and typed, but not too much; very little in fact.
 
 ---
 ## Cites
 
-#### There are many repositories demonstrating data structures in JavaScript. 
+#### There are many repositories demonstrating data structures in JavaScript.
 
 Here are a few: [datastructures-js](https://github.com/eyas-ranjous/datastructures-js/) by [Eyas Ranjous](https://github.com/eyas-ranjous), [JS Shelf](https://github.com/js-shelf), [Itsy Bitsy Data Structures](https://github.com/jamiebuilds/itsy-bitsy-data-structures/) by [James Kyle](https://github.com/jamiebuilds), [Bucke-JS](https://github.com/mauriciosantos/Buckets-JS) by [Mauricio Santos](https://github.com/mauriciosantos/Buckets-JS).
 
-#### There are many online classes dedicated to teaching data structures in JavaScript. 
+#### There are many online classes dedicated to teaching data structures in JavaScript.
 
 Here are a few: [Kyle Shevlin](https://github.com/kyleshevlin)'s [data structures](https://egghead.io/courses/data-structures-and-algorithms-in-javascript) class on Egg Head, [Algo Expert](algoexpert.io) is a tool for testing algorithm and data structure knowledge!
