+++ package.json
@@ -7,7 +7,7 @@
     "packages/*"
   ],
   "scripts": {
-    "build": "pnpm -r build",
+    "build": "turbo run build",
     "clean": "pnpm -r clean",
     "commit": "git-cz",
     "commit-msg": "commitlint --edit $1",
@@ -18,7 +18,7 @@
     "lint": "eslint . --ext '.ts,' --cache --fix",
     "lint:ci": "eslint . --ext '.ts,' --cache",
     "test": "vitest run --coverage",
-    "test:ci": "pnpm test --ci",
+    "test:ci": "pnpm run test",
     "tsc": "tsc -p ./tsconfig.test.json",
     "pre-commit": "cross-env FORCE_COLOR=1 lint-staged",
     "prepare": "is-ci || pnpm husky-setup",
@@ -44,12 +44,11 @@
     "@commitlint/cli": "^17.0.3",
     "@commitlint/config-lerna-scopes": "^17.0.2",
     "@commitlint/prompt": "^17.0.3",
-    "@heartly/eslint-config-typescript": "^0.0.2",
-    "@types/jest": "^28.1.4",
     "@types/node": "^18.0.0",
     "@typescript-eslint/eslint-plugin": "^5.30.3",
     "@typescript-eslint/parser": "^5.30.3",
-    "c8": "^7.11.3",
+    "@vitest/coverage-c8": "^0.26.3",
+    "c8": "^7.12.0",
     "changesets": "^1.0.2",
     "commitizen": "^4.0.4",
     "conventional-changelog-cli": "^2.0.11",
@@ -63,8 +62,9 @@
     "rimraf": "^3.0.2",
     "tslib": "^2.4.0",
     "tsutils": "^3.21.0",
+    "turbo": "^1.6.3",
     "typescript": "^4.7.4",
-    "vitest": "^0.16.0"
+    "vitest": "^0.26.3"
   },
   "config": {
     "commitizen": {
