+++ packages/linked-list/tsconfig.json
@@ -3,6 +3,5 @@
   "compilerOptions": {
     "outDir": "dist",
   },
-  "include": ["src"]
+  "include": ["src", "tests"]
 }
-
