+++ packages/queue/tsconfig.json
@@ -3,5 +3,5 @@
   "compilerOptions": {
     "outDir": "dist",
   },
-  "include": ["src"]
+  "include": ["src", "tests"]
 }
